Restructure teachyourselfcs directory
Im struggling just a bit more with some of the exercise in SICP than I care to
admit. By restructuring this a bit, I can make room for multiple "books" in each
subject from teachyourselfcs.com. Part of this means that I plan to revist the
rest of SICP in the future